oops need to typecast metasocket

+++ Core/HLE/sceNetAdhoc.cpp
@@ -336,8 +336,8 @@ static void __AdhocctlNotify(u64 userdata, int cyclesLate) {
 			ret = SOCKET_ERROR;
 			sockerr = EAGAIN;
 			// Don't send anything yet if connection to Adhoc Server is still in progress
-			if (!isAdhocctlNeedLogin && IsSocketReady(metasocket, false, true) > 0) {
-				ret = send(metasocket, (const char*)&packet, len, MSG_NOSIGNAL);
+			if (!isAdhocctlNeedLogin && IsSocketReady((int)metasocket, false, true) > 0) {
+				ret = send((int)metasocket, (const char*)&packet, len, MSG_NOSIGNAL);
 				sockerr = errno;
 				// Successfully Sent or Connection has been closed or Connection failure occurred
 				if (ret >= 0 || (ret == SOCKET_ERROR && sockerr != EAGAIN && sockerr != EWOULDBLOCK)) {
@@ -403,7 +403,7 @@ static void __AdhocctlState(u64 userdata, int cyclesLate) {
 
 // Used to simulate blocking on metasocket when send OP code to AdhocServer
 int WaitBlockingAdhocctlSocket(AdhocctlRequest request, int usec, const char* reason) {
-	int uid = (metasocket <= 0) ? 1 : metasocket;
+	int uid = (metasocket <= 0) ? 1 : (int)metasocket;
 
 	if (adhocctlRequests.find(uid) != adhocctlRequests.end()) {
 		WARN_LOG(SCENET, "sceNetAdhocctl - WaitID[%d] already existed, Socket is busy!", uid);
@@ -2383,7 +2383,7 @@ u32 NetAdhocctl_Disconnect() {
 			//_acquireNetworkLock();
 
 			// Send Disconnect Request Packet
-			iResult = send(metasocket, (const char*)&opcode, 1, MSG_NOSIGNAL);
+			iResult = send((int)metasocket, (const char*)&opcode, 1, MSG_NOSIGNAL);
 			error = errno;
 
 			// Sending may get socket error 10053 if the AdhocServer is already shutted down
@@ -2500,8 +2500,8 @@ int NetAdhocctl_Term() {
 		adhocctlHandlers.clear();
 		// Free stuff here
 		networkInited = false;
-		shutdown(metasocket, SD_BOTH);
-		closesocket(metasocket);
+		shutdown((int)metasocket, SD_BOTH);
+		closesocket((int)metasocket);
 		metasocket = (int)INVALID_SOCKET;
 		// Delete fake PSP Thread. 
 		// kernelObjects may already been cleared early during a Shutdown, thus trying to access it may generates Warning/Error in the log
